libexec/counttests.sh: Restrict overhead
counttests.sh calls 'make all_check' before counting the number of
scheduled tests to ensure that test executables and generated test
YAML files are present. In its current implementation this call is not
restricted to the tests specified on the command line, resulting in
compilation of unnecessary test files.

Fix this by ensuring that 'make all_check' is only called for the
specified tests.

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>