tela: Fix handling of warnings with BEFORE/AFTER
During a 'make check BEFORE=.../AFTER=...' invocation, warning messages
such as 'Resource unavailable' that are issued during the resource
matching stage are not logged to the resulting TAP file. This is
inconsistent from how warning messages in other stages of invocation are
handled.

Fix this by ensuring that all warnings emitted in that stage are
logged as TAP diagnostics data.

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>
Reported-by: Alexander Egorenkov <egorenar@linux.ibm.com>